client: add read-ignore parameter
prometheus sends us a query with an end value equal to now.
It means we are quering graphite for each query on prometheus
and then merging local and graphite responses.

While waiting for a proper prometheus fix, we should prevent
remote-read on graphite for queries on the last minutes/hours.

this patch implements the possibility to ignore a timerange from
prometheus requests.
If the new `until` value is newer than `from` we return an empty
response.

Signed-off-by: William Dauchy <w.dauchy@criteo.com>