graphite: add a way to periodically reset the carbon connection
This is important when using UDP because you might not notice
that the host has now moved to another IP. This way the DNS gets
re-resolved and the connection reset to where it should be. Using
TCP here might not always be doable. This can be disabled by
setting a very large duration but should never hurt.

See #23 and https://github.com/prometheus/prometheus/pull/3635