Fix issue #57
This patch fix issue #57--Can not set --read.delay=0s.
It is because the reload function which in main.go will merge overwritting cliCfg into cfg which is a DefaultConfig.
The merge func recognises  '0s' as a empty value and will not overwite it into cfg. As a result, the read.delay will be '1h' which already set in DefaultConfig.
Same problem happens when set 'write.timeout=0s' or 'read.timeout=0s'.